requirements.txt
# A2A Supply Chain Optimization - MVP Requirements
# Python 3.11+

# Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
alembic==1.13.1

# Redis (Phase 2)
# redis==5.0.1

# Data Processing
pandas==2.1.4
numpy==1.26.3

# Machine Learning
scikit-learn==1.4.0
scipy==1.11.4

# Async
aiohttp==3.9.1
asyncio==3.4.3

# Logging
python-json-logger==2.0.7

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0

# Development
black==23.12.1
isort==5.13.2
flake8==7.0.0
mypy==1.8.0

# Environment
python-dotenv==1.0.0
